moved api/users/route.js into app  and added getemail which adds email on the supabase
